float from -1 to 1 for beat strength

+++ src/cpulse.h
@@ -1,3 +1,3 @@
 void cpulse_start(void);
-int * cpulse_pulse(void);
+float * cpulse_pulse(void);
 void cpulse_stop(void);
